pyproject.toml
[project]
name = "gds-idea-ragchat-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==0.2.10",
    "langchain-aws>=0.1.18",
    "langchain-community>=0.2.10",
    "opensearch-py>=2.8.0",
    "pypdf>=5.4.0",
    "requests-aws4auth>=1.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_chat_agent"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=rag_chat_agent"

[dependency-groups]
dev = [
    "boto3>=1.34.162",
    "python-dotenv>=1.1.0",
]

